requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cors-headers==4.3.1
django-mssql-backend==2.8.1
djangorestframework==3.14.0
mssql-django==1.3
mysqlclient==2.2.1
PyJWT==2.8.0
pyodbc==5.0.1
pytz==2023.3.post1
# pywin32==306
random-password-generator==2.2.0
sqlparse==0.4.4
tzdata==2023.4
python-dateutil
